Fix apps with items on dexie
Move code that interacts with the db into db.ts where appropriate.
